# trie: simplify StackTrie implementation