Trim the search key from head as it's being pushed deeper into the trie.
Previously the search key was never modified but each node kept
information how to slice and compare it in `keyOffset`. Now the `keyOffset`
is not needed as this information is included in the slice of the
search key. This way the `keyOffset` can be removed and key manipulation
simplified.

BTW, I ported this Trie implementation (without hashing implementation) to C++. This helped me a lot, thanks.