@holiman 2021-11-22
The benchmarks don't seem affected -- so sure, simplicity is better. 
```
[user@work trie]$ benchstat before.txt after.txt 
name                             old time/op    new time/op    delta
HexToCompact-6                     17.2ns ± 1%    17.8ns ± 5%   ~     (p=0.400 n=3+3)
CompactToHex-6                     31.2ns ± 1%    31.6ns ± 1%   ~     (p=0.200 n=3+3)
KeybytesToHex-6                    33.6ns ± 4%    33.9ns ± 3%   ~     (p=1.000 n=3+3)
HexToKeybytes-6                    18.9ns ± 0%    19.0ns ± 0%   ~     (p=0.100 n=3+3)
Prove-6                             397µs ± 5%     457µs ± 4%   ~     (p=0.100 n=3+3)
VerifyProof-6                      6.44µs ± 3%    6.10µs ± 5%   ~     (p=0.200 n=3+3)
VerifyRangeProof10-6               64.4µs ± 1%    61.0µs ± 3%   ~     (p=0.100 n=3+3)
VerifyRangeProof100-6               342µs ± 4%     336µs ± 6%   ~     (p=0.400 n=3+3)
VerifyRangeProof1000-6             3.20ms ±15%    3.05ms ± 9%   ~     (p=0.700 n=3+3)
VerifyRangeProof5000-6             11.0ms ± 2%    10.7ms ± 2%   ~     (p=0.200 n=3+3)
VerifyRangeNoProof10-6              312µs ± 4%     303µs ± 7%   ~     (p=0.400 n=3+3)
VerifyRangeNoProof500-6            1.25ms ± 5%    1.24ms ± 2%   ~     (p=1.000 n=3+3)
VerifyRangeNoProof1000-6           2.32ms ± 0%    2.36ms ± 1%   ~     (p=0.200 n=3+3)
Get-6                               187ns ± 4%     182ns ± 1%   ~     (p=0.400 n=3+3)
GetDB-6                             163ns ± 1%     157ns ± 1%   ~     (p=0.100 n=3+3)
UpdateBE-6                         1.30µs ± 3%    1.39µs ± 7%   ~     (p=0.400 n=3+3)
UpdateLE-6                         1.86µs ± 5%    1.88µs ± 6%   ~     (p=1.000 n=3+3)
Hash-6                             2.12µs ±17%    1.95µs ± 7%   ~     (p=0.400 n=3+3)
CommitAfterHash/no-onleaf-6        2.15µs ± 5%    2.09µs ± 6%   ~     (p=1.000 n=3+3)
CommitAfterHash/with-onleaf-6      2.49µs ± 8%    2.41µs ± 3%   ~     (p=0.700 n=3+3)
HashFixedSize/10-6                 54.9µs ± 4%    52.0µs ± 4%   ~     (p=0.200 n=3+3)
HashFixedSize/100-6                 223µs ± 7%     207µs ± 4%   ~     (p=0.100 n=3+3)
HashFixedSize/1K-6                 1.51ms ±12%    1.30ms ± 2%   ~     (p=0.100 n=3+3)
HashFixedSize/10K-6                12.9ms ±13%    10.7ms ± 1%   ~     (p=0.100 n=3+3)
HashFixedSize/100K-6                124ms ±20%     104ms ± 4%   ~     (p=0.100 n=3+3)
CommitAfterHashFixedSize/10-6      33.3µs ±17%    27.8µs ± 2%   ~     (p=0.100 n=3+3)
CommitAfterHashFixedSize/100-6      138µs ± 1%     132µs ± 1%   ~     (p=0.100 n=3+3)
CommitAfterHashFixedSize/1K-6      1.34ms ± 5%    1.23ms ± 2%   ~     (p=0.100 n=3+3)
CommitAfterHashFixedSize/10K-6     17.0ms ± 5%    17.2ms ± 5%   ~     (p=1.000 n=3+3)
CommitAfterHashFixedSize/100K-6     194ms ± 5%     179ms ± 4%   ~     (p=0.200 n=3+3)
DerefRootFixedSize/10-6            15.3µs ± 2%    14.4µs ± 1%   ~     (p=0.100 n=3+3)
DerefRootFixedSize/100-6           48.3µs ± 5%    45.8µs ± 0%   ~     (p=0.100 n=3+3)
DerefRootFixedSize/1K-6             376µs ± 5%     362µs ± 5%   ~     (p=0.400 n=3+3)
DerefRootFixedSize/10K-6           4.47ms ± 3%    4.31ms ± 1%   ~     (p=0.200 n=3+3)
DerefRootFixedSize/100K-6          51.8ms ± 6%    49.5ms ± 2%   ~     (p=0.400 n=3+3)

name                             old alloc/op   new alloc/op   delta
HexToCompact-6                      4.00B ± 0%     4.00B ± 0%   ~     (all equal)
CompactToHex-6                      16.0B ± 0%     16.0B ± 0%   ~     (all equal)
KeybytesToHex-6                     24.0B ± 0%     24.0B ± 0%   ~     (all equal)
HexToKeybytes-6                     4.00B ± 0%     4.00B ± 0%   ~     (all equal)
Prove-6                             136kB ± 2%     154kB ± 2%   ~     (p=0.100 n=3+3)
VerifyProof-6                      5.09kB ± 5%    5.16kB ± 4%   ~     (p=0.800 n=3+3)
VerifyRangeProof10-6               42.6kB ± 0%    42.5kB ± 0%   ~     (p=0.200 n=3+3)
VerifyRangeProof100-6               270kB ± 0%     271kB ± 0%   ~     (p=1.000 n=3+3)
VerifyRangeProof1000-6             2.06MB ± 1%    2.07MB ± 1%   ~     (p=1.000 n=3+3)
VerifyRangeProof5000-6             10.0MB ± 0%    10.0MB ± 0%   ~     (p=0.400 n=3+3)
VerifyRangeNoProof10-6             57.0kB ± 0%    55.0kB ± 1%   ~     (p=0.100 n=3+3)
VerifyRangeNoProof500-6             200kB ± 2%     199kB ± 2%   ~     (p=1.000 n=3+3)
VerifyRangeNoProof1000-6            366kB ± 1%     365kB ± 0%   ~     (p=0.700 n=3+3)
Get-6                                104B ± 0%      104B ± 0%   ~     (all equal)
GetDB-6                              104B ± 0%      104B ± 0%   ~     (all equal)
UpdateBE-6                         1.80kB ± 0%    1.79kB ± 0%   ~     (p=1.000 n=3+3)
UpdateLE-6                         1.85kB ± 0%    1.85kB ± 0%   ~     (p=1.000 n=3+3)
Hash-6                               952B ± 1%      958B ± 1%   ~     (p=0.400 n=3+3)
CommitAfterHash/no-onleaf-6          818B ± 3%      806B ± 2%   ~     (p=0.700 n=3+3)
CommitAfterHash/with-onleaf-6        983B ± 8%     1011B ± 1%   ~     (p=1.000 n=3+3)
HashFixedSize/10-6                 12.0kB ± 0%    12.0kB ± 0%   ~     (p=0.700 n=3+3)
HashFixedSize/100-6                59.6kB ± 0%    59.6kB ± 0%   ~     (p=0.700 n=3+3)
HashFixedSize/1K-6                  593kB ± 0%     593kB ± 0%   ~     (p=0.100 n=3+3)
HashFixedSize/10K-6                6.31MB ± 0%    6.31MB ± 0%   ~     (p=0.200 n=3+3)
HashFixedSize/100K-6               62.4MB ± 0%    62.4MB ± 0%   ~     (p=0.100 n=3+3)
CommitAfterHashFixedSize/10-6      16.6kB ± 0%    16.6kB ± 0%   ~     (p=0.800 n=3+3)
CommitAfterHashFixedSize/100-6     80.4kB ± 0%    80.4kB ± 0%   ~     (p=0.800 n=3+3)
CommitAfterHashFixedSize/1K-6       763kB ± 0%     763kB ± 0%   ~     (p=0.700 n=3+3)
CommitAfterHashFixedSize/10K-6     9.04MB ± 0%    9.04MB ± 0%   ~     (p=1.000 n=3+3)
CommitAfterHashFixedSize/100K-6    83.9MB ± 0%    83.9MB ± 0%   ~     (p=0.700 n=3+3)
DerefRootFixedSize/10-6              745B ± 0%      745B ± 0%   ~     (all equal)
DerefRootFixedSize/100-6             744B ± 0%      744B ± 0%   ~     (all equal)
DerefRootFixedSize/1K-6              760B ± 0%      760B ± 0%   ~     (all equal)
DerefRootFixedSize/10K-6             760B ± 0%      760B ± 0%   ~     (all equal)
DerefRootFixedSize/100K-6            760B ± 0%      760B ± 0%   ~     (all equal)

name                             old allocs/op  new allocs/op  delta
HexToCompact-6                       1.00 ± 0%      1.00 ± 0%   ~     (all equal)
CompactToHex-6                       1.00 ± 0%      1.00 ± 0%   ~     (all equal)
KeybytesToHex-6                      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
HexToKeybytes-6                      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Prove-6                             1.86k ± 1%     2.10k ± 1%   ~     (p=0.100 n=3+3)
VerifyProof-6                         104 ± 4%       106 ± 4%   ~     (p=0.700 n=3+3)
VerifyRangeProof10-6                  432 ± 0%       431 ± 0%   ~     (p=0.700 n=3+3)
VerifyRangeProof100-6               1.95k ± 0%     1.95k ± 0%   ~     (p=1.000 n=3+3)
VerifyRangeProof1000-6              16.1k ± 1%     16.1k ± 1%   ~     (p=0.700 n=3+3)
VerifyRangeProof5000-6              79.2k ± 0%     79.0k ± 0%   ~     (p=0.700 n=3+3)
VerifyRangeNoProof10-6              1.20k ± 0%     1.18k ± 0%   ~     (p=0.100 n=3+3)
VerifyRangeNoProof500-6             3.72k ± 1%     3.71k ± 1%   ~     (p=0.700 n=3+3)
VerifyRangeNoProof1000-6            6.75k ± 0%     6.73k ± 0%   ~     (p=0.400 n=3+3)
Get-6                                2.00 ± 0%      2.00 ± 0%   ~     (all equal)
GetDB-6                              2.00 ± 0%      2.00 ± 0%   ~     (all equal)
UpdateBE-6                           9.00 ± 0%      9.00 ± 0%   ~     (all equal)
UpdateLE-6                           9.00 ± 0%      9.00 ± 0%   ~     (all equal)
Hash-6                               11.0 ± 0%      11.0 ± 0%   ~     (all equal)
CommitAfterHash/no-onleaf-6          7.00 ± 0%      7.00 ± 0%   ~     (all equal)
CommitAfterHash/with-onleaf-6        11.0 ± 0%      11.0 ± 0%   ~     (all equal)
HashFixedSize/10-6                    149 ± 0%       149 ± 0%   ~     (all equal)
HashFixedSize/100-6                   770 ± 0%       770 ± 0%   ~     (all equal)
HashFixedSize/1K-6                  7.45k ± 0%     7.45k ± 0%   ~     (p=1.000 n=3+3)
HashFixedSize/10K-6                 77.1k ± 0%     77.1k ± 0%   ~     (p=0.400 n=3+3)
HashFixedSize/100K-6                 768k ± 0%      768k ± 0%   ~     (p=0.700 n=3+3)
CommitAfterHashFixedSize/10-6         162 ± 0%       162 ± 0%   ~     (all equal)
CommitAfterHashFixedSize/100-6        784 ± 0%       784 ± 0%   ~     (all equal)
CommitAfterHashFixedSize/1K-6       7.76k ± 0%     7.76k ± 0%   ~     (all equal)
CommitAfterHashFixedSize/10K-6      80.0k ± 0%     80.0k ± 0%   ~     (p=0.800 n=3+3)
CommitAfterHashFixedSize/100K-6      799k ± 0%      799k ± 0%   ~     (p=0.500 n=3+3)
DerefRootFixedSize/10-6              8.00 ± 0%      8.00 ± 0%   ~     (all equal)
DerefRootFixedSize/100-6             8.00 ± 0%      8.00 ± 0%   ~     (all equal)
DerefRootFixedSize/1K-6              10.0 ± 0%      10.0 ± 0%   ~     (all equal)
DerefRootFixedSize/10K-6             10.0 ± 0%      10.0 ± 0%   ~     (all equal)
DerefRootFixedSize/100K-6            10.0 ± 0%      10.0 ± 0%   ~     (all equal)
```